+++ views/motion/_view_motiondata.php
@@ -1,16 +1,12 @@
 <?php
 
-use app\components\HTMLTools;
-use app\components\Tools;
-use app\components\UrlHelper;
-use app\models\db\Motion;
-use app\models\db\MotionSupporter;
-use app\models\db\User;
+use app\components\{HTMLTools, Tools, UrlHelper};
+use app\models\db\{Motion, MotionSupporter, User};
 use yii\helpers\Html;
 use app\views\motion\LayoutHelper as MotionLayoutHelper;
 
 /**
- * @var \yii\web\View $this
+ * @var Yii\web\View $this
  * @var Motion $motion
  * @var int[] $openedComments
  * @var null|string $supportStatus
@@ -24,7 +20,7 @@ echo '<div class="content">';
 $replacedByMotions = $motion->getVisibleReplacedByMotions();
 if (count($replacedByMotions) > 0) {
     echo '<div class="alert alert-danger motionReplayedBy" role="alert">';
-    echo \Yii::t('motion', 'replaced_by_hint');
+    echo Yii::t('motion', 'replaced_by_hint');
     if (count($replacedByMotions) > 1) {
         echo '<ul>';
         foreach ($replacedByMotions as $newMotion) {
@@ -50,7 +46,7 @@ $motionData[] = [
 
 if ($motion->agendaItem) {
     $motionData[] = [
-        'title'   => \Yii::t('motion', 'agenda_item'),
+        'title'   => Yii::t('motion', 'agenda_item'),
         'content' => Html::encode($motion->agendaItem->getShownCode(true) . ' ' . $motion->agendaItem->title),
     ];
 }
@@ -66,7 +62,7 @@ if (count($initiators) > 0 && !$motion->isResolution()) {
 
 $motionData[] = [
     'rowClass' => 'statusRow',
-    'title'    => \Yii::t('motion', 'status'),
+    'title'    => Yii::t('motion', 'status'),
     'content'  => $motion->getFormattedStatus(),
 ];
 
@@ -105,7 +101,7 @@ if (!$motion->isResolution()) {
     if (($motion->isProposalPublic() && $motion->proposalStatus) || $proposalAdmin) {
         $motionData[] = [
             'rowClass' => 'proposedStatusRow',
-            'title'    => \Yii::t('amend', 'proposed_status'),
+            'title'    => Yii::t('amend', 'proposed_status'),
             'tdClass'  => 'str',
             'content'  => $motion->getFormattedProposalStatus(true),
         ];
@@ -126,14 +122,14 @@ if (count($initiators) > 0 && $motion->isResolution()) {
 
 if ($motion->dateResolution) {
     $motionData[] = [
-        'title'   => \Yii::t('motion', 'resoluted_on'),
+        'title'   => Yii::t('motion', 'resoluted_on'),
         'content' => Tools::formatMysqlDate($motion->dateResolution, null, false),
     ];
 }
 
 if (!$motion->isResolution()) {
     $motionData[] = [
-        'title'   => \Yii::t('motion', ($motion->isSubmitted() ? 'submitted_on' : 'created_on')),
+        'title'   => Yii::t('motion', ($motion->isSubmitted() ? 'submitted_on' : 'created_on')),
         'content' => Tools::formatMysqlDateTime($motion->dateCreation, null, false),
     ];
 }
@@ -148,15 +144,15 @@ if ($admin && count($motion->getMyConsultation()->tags) > 0) {
         $str            = Html::encode($tag->title);
         $str            .= Html::beginForm('', 'post', ['class' => 'form-inline delTagForm delTag' . $tag->id]);
         $str            .= '<input type="hidden" name="tagId" value="' . $tag->id . '">';
-        $str            .= '<button type="submit" name="motionDelTag">' . \Yii::t('motion', 'tag_del') . '</button>';
+        $str            .= '<button type="submit" name="motionDelTag">' . Yii::t('motion', 'tag_del') . '</button>';
         $str            .= Html::endForm();
         $tags[]         = $str;
     }
     $content = implode(', ', $tags);
 
-    $content .= '&nbsp; &nbsp; <a href="#" class="tagAdderHolder">' . \Yii::t('motion', 'tag_new') . '</a>';
+    $content .= '&nbsp; &nbsp; <a href="#" class="tagAdderHolder">' . Yii::t('motion', 'tag_new') . '</a>';
     $content .= Html::beginForm('', 'post', ['id' => 'tagAdderForm', 'class' => 'form-inline hidden']);
-    $content .= '<select name="tagId" title="' . \Yii::t('motion', 'tag_select') . '" class="form-control">
+    $content .= '<select name="tagId" title="' . Yii::t('motion', 'tag_select') . '" class="form-control">
         <option>-</option>';
 
     foreach ($motion->getMyConsultation()->tags as $tag) {
@@ -166,12 +162,12 @@ if ($admin && count($motion->getMyConsultation()->tags) > 0) {
     }
     $content .= '</select>
             <button class="btn btn-primary" type="submit" name="motionAddTag">' .
-                \Yii::t('motion', 'tag_add') .
+                Yii::t('motion', 'tag_add') .
                 '</button>';
     $content .= Html::endForm();
 
     $motionData[] = [
-        'title'   => \Yii::t('motion', 'tag_tags'),
+        'title'   => Yii::t('motion', 'tag_tags'),
         'tdClass' => 'tags',
         'content' => $content,
     ];
@@ -181,7 +177,7 @@ if ($admin && count($motion->getMyConsultation()->tags) > 0) {
         $tags[] = $tag->title;
     }
     $motionData[] = [
-        'title'   => (count($motion->tags) > 1 ? \Yii::t('motion', 'tags') : \Yii::t('motion', 'tag')),
+        'title'   => (count($motion->tags) > 1 ? Yii::t('motion', 'tags') : Yii::t('motion', 'tag')),
         'content' => Html::encode(implode(', ', $tags)),
     ];
 }
@@ -193,7 +189,7 @@ if ($motion->replacedMotion) {
     $changesLink = UrlHelper::createMotionUrl($motion, 'view-changes');
     $content     .= '<div class="changesLink">';
     $content     .= '<span class="glyphicon glyphicon-chevron-right"></span> ';
-    $content     .= Html::a(\Yii::t('motion', 'replaces_motion_diff'), $changesLink);
+    $content     .= Html::a(Yii::t('motion', 'replaces_motion_diff'), $changesLink);
     $content     .= '</div>';
 
     if ($motion->isResolution() && !$motion->replacedMotion->isResolution()) {
@@ -213,8 +209,8 @@ if ((!isset($skip_drafts) || !$skip_drafts) && $motion->getMergingDraft(true)) {
     $url          = UrlHelper::createMotionUrl($motion, 'merge-amendments-public');
     $motionData[] = [
         'rowClass' => 'mergingDraft',
-        'title'    => \Yii::t('motion', 'merging_draft_th'),
-        'content'  => str_replace('%URL%', Html::encode($url), \Yii::t('motion', 'merging_draft_td')),
+        'title'    => Yii::t('motion', 'merging_draft_th'),
+        'content'  => str_replace('%URL%', Html::encode($url), Yii::t('motion', 'merging_draft_td')),
     ];
 }
 
@@ -231,7 +227,7 @@ if (User::getCurrentUser()) {
         $str .= HTMLTools::textToHtmlWithLink($comment ? $comment->text : '') . '</blockquote>';
     }
     $str .= Html::beginForm('', 'post', ['class' => 'form-inline' . ($comment ? ' hidden' : '')]);
-    $str .= '<textarea class="form-control" name="noteText" title="' . \Yii::t('motion', 'private_notes') . '">';
+    $str .= '<textarea class="form-control" name="noteText" title="' . Yii::t('motion', 'private_notes') . '">';
     if ($comment) {
         $str .= Html::encode($comment->text);
     }
@@ -239,12 +235,12 @@ if (User::getCurrentUser()) {
     $str .= '<input type="hidden" name="paragraphNo" value="-1">';
     $str .= '<input type="hidden" name="sectionId" value="">';
     $str .= '<button type="submit" name="savePrivateNote" class="btn btn-success">' .
-            \Yii::t('base', 'save') . '</button>';
+            Yii::t('base', 'save') . '</button>';
     $str .= Html::endForm();
 
     $motionData[] = [
         'rowClass' => 'privateNotes' . ($comment ? '' : ' hidden'),
-        'title'    => \Yii::t('motion', 'private_notes'),
+        'title'    => Yii::t('motion', 'private_notes'),
         'content'  => $str,
     ];
 }