+++ web/typescript/Antragsgruen.ts
@@ -71,16 +71,18 @@ declare let ANTRAGSGRUEN_STRINGS: string[][];
     });
 
     // Needs to be synchronized with ConsultationAgendaItem:getSortedFromConsultation
-    let recalcAgendaNode = function ($ol) {
-        let currNumber = '0.',
+    const recalcAgendaNode = ($ol: JQuery, prefix: string) => {
+        const separator = '.';
+        const prevCode = (prefix === '' || prefix[prefix.length - 1] === separator ? prefix : prefix + separator);
+        let currNumber = '0' + separator,
             $lis = $ol.find('> li.agendaItem');
         $lis.each(function () {
             let $li = $(this),
                 code = $li.data('code'),
-                currStr = '',
+                currStr,
                 $subitems = $li.find('> ol');
             if (code == '#') {
-                let parts = currNumber.split('.');
+                let parts = currNumber.split(separator);
                 if (parts[0].match(/^[a-y]$/i)) { // Single alphabetical characters
                     parts[0] = String.fromCharCode(parts[0].charCodeAt(0) + 1);
                 } else { // Numbers or mixtures of alphabetical characters and numbers
@@ -89,19 +91,22 @@ declare let ANTRAGSGRUEN_STRINGS: string[][];
                         numeric = parseInt(matches[2] == '' ? '1' : matches[2]);
                     parts[0] = nonNumeric + ++numeric;
                 }
-                currNumber = currStr = parts.join('.');
+                currNumber = currStr = parts.join(separator);
             } else {
                 currStr = currNumber = code + ''; // currNumber needs to be a string, always.
+            }
+            if (currStr !== '') {
+                currStr = prevCode + currStr;
             }
 
             $li.find('> div > h3 .code').text(currStr);
             if ($subitems.length > 0) {
-                recalcAgendaNode($subitems);
+                recalcAgendaNode($subitems, currStr);
             }
         });
     };
     $('ol.motionListWithinAgenda').on("antragsgruen:agenda-change", function () {
-        recalcAgendaNode($(this));
+        recalcAgendaNode($(this), '');
     }).trigger("antragsgruen:agenda-change");
 
     $('.motionList .amendmentsToggler').each((i, el) => {